# docs/swagger.yaml
definitions:
...
  models.Graph:
    properties:
      areaName:
        type: string
      gradeCode:
        type: string
      graphLine:
        items:
          $ref: '#/definitions/models.GraphLine'
        type: array
      itemCode:
        type: integer
      itemKindCode:
        type: string
      lastShipDate:
        type: string
      rangeLabel:
        items:
          type: string
        type: array
      rangeMax:
        type: integer
      rangeMin:
        type: integer
      rangeStep:
        type: integer
      rangeType:
        type: string
    type: object
  models.GraphData:
    properties:
      x:
        type: string
      "y":
        type: number
    type: object
  models.GraphLine:
    properties:
      areaName:
        type: string
      graphData:
        items:
          $ref: '#/definitions/models.GraphData'
        type: array
      marketName:
        type: string
    type: object
  models.Item:
    properties:
      itemCode:
        type: integer
      itemName:
        type: string
    type: object
  models.ItemGrade:
    properties:
      gradeCode:
        type: string
      gradeName:
        type: string
      gradeOrder:
        type: integer
      gradeType:
        type: integer
      shipAmt:
        type: number
      shipUnit:
        type: string
    type: object
  models.ItemKind:
    properties:
      itemCode:
        type: integer
      itemKindCode:
        type: string
      itemKindName:
        type: string
      itemName:
        type: string
    type: object
  models.MonthInfo:
    properties:
      areaName:
        type: string
      baseMonth:
        type: string
      gradeCode:
        type: string
      itemCode:
        type: integer
      itemKindCode:
        type: string
      marketName:
        type: string
      monthPriceList:
        items:
          $ref: '#/definitions/models.MonthPrice'
        type: array
      shipMonth:
        type: integer
      shipYear:
        type: integer
    type: object
  models.MonthPrice:
    properties:
      avgPrice:
        type: number
      baseMonth:
        type: string
      maxPrice:
        type: integer
      maxShipDate:
        type: string
      minPrice:
        type: integer
      minShipDate:
        type: string
      priceCount:
        type: integer
      sumPrice:
        type: integer
    type: object
...
  models.RecentPrice:
    properties:
      areaName:
        type: string
      gradeCode:
        type: string
      itemCode:
        type: integer
      itemKindCode:
        type: string
      lastPrice:
        type: integer
      lastShipDate:
        type: string
      marketName:
        type: string
      stat7days:
        $ref: '#/definitions/models.StatPrice'
        type: object
      stat30days:
        $ref: '#/definitions/models.StatPrice'
        type: object
    type: object
  models.StatPrice:
    properties:
      avgPrice:
        type: number
      cntPrice:
        type: integer
      days:
        type: integer
      maxPrice:
        type: integer
      minPrice:
        type: integer
      sumPrice:
        type: integer
    type: object
...
  v2.WholePrice:
    properties:
      examinDate:
        type: string
      examinGradeCode:
        type: string
      examinGradeName:
        type: string
      examinItemCode:
        type: string
      examinItemName:
        type: string
      examinSpeciesCode:
        type: string
      examinSpeciesName:
        type: string
      examinUnit:
        type: string
      examinUnitName:
        type: string
      price:
        type: integer
    type: object
  v2.WholePriceGraph:
    properties:
      graphLine:
        items:
          $ref: '#/definitions/v2.wholePriceGraphLine'
        type: array
      rangeLabel:
        items:
          type: string
        type: array
      stdItemCode:
        type: string
    type: object
...
  v2.wholePriceGraphData:
    properties:
      x:
        type: string
      "y":
        type: integer
    type: object
  v2.wholePriceGraphLine:
    properties:
      examinGradeName:
        type: string
      examinItemName:
        type: string
      examinSpeciesName:
        type: string
      examinUnitName:
        type: string
      graphData:
        items:
          $ref: '#/definitions/v2.wholePriceGraphData'
        type: array
    type: object
info:
  contact: {}
  description: 아그리파 API
  license: {}
  title: AGRIPA API
  version: "1.0"
paths:
...
  /v2/whole/price/line-graph/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 도매가격 꺾은선 그래프 정보 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.WholePriceGraph'
      summary: 도매가격 꺾은선 그래프 정보 조회
      tags:
      - v2 도매
  /v2/whole/price/previous-year/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 전년 도매가격 정보 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.WholePrice'
      summary: 전년 도매가격 정보 조회
      tags:
      - v2 도매
  /v2/whole/price/recent/{std-item-code}:
    get:
      consumes:
      - application/json
      description: 최근 도매가격 정보 조회 조회합니다
      parameters:
      - description: 표준품목코드
        in: path
        name: std-item-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.WholePrice'
      summary: 최근 도매가격 정보 조회
      tags:
      - v2 도매
...
  /whole/grade/{itemcode}/{kindcode}:
    get:
      consumes:
      - application/json
      description: 등급 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemGrade'
      summary: 등급 정보 조회
      tags:
      - 도매
  /whole/item/{catecode}:
    get:
      consumes:
      - application/json
      description: 품목 정보를 조회합니다
      parameters:
      - description: 부류코드
        in: path
        name: catecode
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Item'
      summary: 품목 정보 조회
      tags:
      - 도매
  /whole/kind/{itemcode}:
    get:
      consumes:
      - application/json
      description: 품종 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemKind'
      summary: 품종 정보 조회
      tags:
      - 도매
  /whole/price/graph/{itemcode}/{kindcode}/{gradecode}/{graphType}:
    get:
      consumes:
      - application/json
      description: 도매 가격 정보 - 그래프용
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      - description: 등급코드
        in: path
        name: gradecode
        required: true
        type: string
      - description: 그래프타입(day,month,year)
        in: path
        name: graphType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Graph'
      summary: 도매 가격 정보(그래프용)
      tags:
      - 도매
  /whole/price/month/{itemcode}/{kindcode}/{gradecode}:
    get:
      consumes:
      - application/json
      description: 도매 전월 기준 3년간 가격 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      - description: 등급코드
        in: path
        name: gradecode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MonthInfo'
      summary: 도매 월 가격 정보 조회
      tags:
      - 도매
  /whole/price/recent/{itemcode}/{kindcode}/{gradecode}:
    get:
      consumes:
      - application/json
      description: 도매 최근 가격 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      - description: 등급코드
        in: path
        name: gradecode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RecentPrice'
      summary: 도매 최근 가격 정보 조회
      tags:
      - 도매
swagger: "2.0"